+++ thepipe/tests/test_provenance.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import tempfile
 import unittest
-from thepipe import Provenance, disable_provenance
+from thepipe import Provenance
 
 
 class TestProvenance(unittest.TestCase):
@@ -91,10 +91,3 @@ class TestProvenance(unittest.TestCase):
         fobj = tempfile.NamedTemporaryFile(delete=True)
         p._export(fobj.name)
         assert open(fobj.name, "r").read() == p.as_json(indent=2)
-
-    def test_outfile_is_not_written_when_disabled(self):
-        p = Provenance()
-        p.reset()
-        assert p.outfile is not None
-        disable_provenance()
-        assert p.outfile is None